# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AJN Luxury Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ORDERS
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Approve/reject review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # INVENTORY
...
components:
  schemas:
...
    ProductList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    Product:
      type: object
      required: [id, nameAr, price, category, inStock]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        price:
          type: number
        discountPrice:
          type: ["number", "null"]
        category:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ProductColor"
        inStock:
          type: boolean
        stockQuantity:
          type: integer
        hasCustomization:
          type: boolean
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductColor:
      type: object
      required: [name, hex]
      properties:
        name:
          type: string
        hex:
          type: string
    ProductInput:
      type: object
      required: [nameAr, price, category]
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        category:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ProductColor"
        stockQuantity:
          type: integer
        hasCustomization:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        category:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ProductColor"
        stockQuantity:
          type: integer
        hasCustomization:
          type: boolean
        inStock:
          type: boolean
...
    Review:
      type: object
      required: [id, productId, customerName, rating, images, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [productId, rating, customerName]
      properties:
        productId:
          type: integer
        customerName:
          type: string
        customerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        images:
          type: array
          items:
            type: string
    ReviewUpdate:
      type: object
      properties:
        status:
          type: string
...