# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AJN Luxury Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductColor } from './productColor';

export interface ProductInput {
  nameAr: string;
  nameEn?: string;
  descriptionAr?: string;
  price: number;
  discountPrice?: number;
  category: string;
  images?: string[];
  colors?: ProductColor[];
  stockQuantity?: number;
  hasCustomization?: boolean;
}